fix: switch setter to use pointers
Before this change all the generated model setter were noop. They set
values to the struct copy not the struct itself and since that copy
was never returned they were noop.

Also, added an update script for replacing expected models in the test
cases but with the great power comes great responsibility, so review
the overwrites carefully.