fix: improve allOf type generation (#48)
When merging allOfs, remove the type reference value _unless_ it is a
self-refernce or is an allOf with exactly one value.

These exceptions are special cases. The end result should cause flat
structs to be generated with the merged fields from each allOf. The
execptions are provided because (1) self-references can not be flattened
and (2) the allOf with one field is a special behavior that is just
overwriting properties on the single model, usually marking it nullable
or adding a description, so this is a nice special case that is better
in the old behavior.

Improve the logic for the exception when an allOf contains a single
reference. This merge now treats it as merging two refs and doesn't need
to do the additional recursion. This ensures the correct references are
returned for for struct properties.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>