feat: improve description handling in allOfs
Add logic to skip semantically empty schemas inside an allOf. For
example, skipping items that only modify the description. It would be
better to bubble up some of this metadata, but this would require a
significant refactor of the object handling. This change should reduce
the number of situations that result in inline structs being generated,
thus improving the resulting code quality. The tradeoff is that the code
comments are a little less good. But the improved types seem to be worth
it.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>