feat: Add discriminator support for oneOfs
Add support for the discriminator property on oneOfs. This allows the
envelope type to implement validation, by using the descriminator to
call validate on the child data. This depends on the fact that we can
now use the discriminator to unmarshal json data into concrete
sub-types that implement validation.

Additionally, using the embed package, all model templates have been
moved into external files that are then embedded into the compiled
binary. This should improve the developer experience.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>