feat: allow setting additional supported string formats
Add new flag that allows us to control the validation that is run after
we filter an api spec. Specifically, it allows us to control the string
formats that will be recognized. The default value adds the `uuid`
format, which we use extensively. But the value can be easily extended
with the new flag.

Additionally, return the filterd yaml string and print it to stdout when
the log level is set to debug, this allows us to then run the command
and if there are validation errors, see the schema value it was trying
to validate.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>