feat: Add discriminator support for oneOfs (#57)
Add support for the discriminator property on oneOfs. This allows the
envelope type to implement validation, by using the descriminator to
call validate on the child data. This depends on the fact that we can
now use the discriminator to unmarshal json data into concrete
sub-types that implement validation.

Additionally, using the embed package, all model templates have been
moved into external files that are then embedded into the compiled
binary. This should improve the developer experience.

Add tests for the serialization of discriminated oneOf


feat: allow setting additional supported string formats

Add new flag that allows us to control the validation that is run after
we filter an api spec. Specifically, it allows us to control the string
formats that will be recognized. The default value adds the `uuid`
format, which we use extensively. But the value can be easily extended
with the new flag.

Additionally, return the filtered yaml string and print it to stdout when
the log level is set to debug, this allows us to then run the command
and if there are validation errors, see the schema value it was trying
to validate.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>