feat: use enum for discriminators
Create an enum for the oneOf discriminator values. This then allows for
linting checks such as 	https://github.com/nishanths/exhaustive to
verify that all cases are handled.

This also improves casting of the discriminator value to match the
documentation in the spec
https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.2.md#discriminatorObject,
specifically about tooling being allowed to convert values to strings.

To verify that `exhaustive` will in-fact cause a linting error if you miss a value in a switch. When you run `exhaustive` on this file:

```go
// check.go
package generatortest

func check() bool {

	foo := Error{}
	switch foo.ErrorDiscriminator() {
	case ErrorDiscriminatorAuth:
		return true
	}
	return false
}
```
It produces

```sh
check.go:6:2: missing cases in switch of type ErrorDiscriminator: ErrorDiscriminatorExternal, ErrorDiscriminatorField, ErrorDiscriminatorGeneric
```

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>