fix: sort model names prior to generation to improve stability (#69)
Improve the generation stability by sorting the models prior to
generating the models. This ensures that we always visit the models in
the same order, which is important because we are iterating over a map
of pointers, which can potentially mutate as we iterate through the
models.

This change makes the codegen deterministic, but is probalby the wrong
order to visit the models in. In a future release, we should consider
some kind of tree structure to ensure we visit in a consistent and
correct order.

fix: always prefer named models as struct props, except for slices (#69)

This change ensures that we return ref names instead of resolved
primitive types if a model has a field that references another model.
With one exception, when the field is a slice, we prefer to resolve the
model so that we avoid models that are slice aliases.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>