feat: add mapstructure tags during generation
In addition to JSON tags, we add the mapstructure tags because these are
needed for oneOf types to behave correctly. We use mapstructure
internally and this will drop data if (1) the name is not camelCase and
(2) this tag is missing.

This is demonstrated by the snake_casing example that is now added to
the test suite.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>